`[mercury]` general fixes (#383)
* Fix vertical alignment of labels in inline fields

Added a margin-block-start to labels positioned on the left or right of input fields for better vertical alignment. This adjustment is applied only when the parent has the field-inline class. Previously, labels positioned above fields inherited the margin-block-start, which caused alignment issues. The solution involves using the selector .field-inline > .label to ensure proper styling.

* update radio-group showcase page

Add chameleon vertical case.

* rename `form-input` to just `input`

an input could be used independently of a form.

* rename `form-label` to just `label`

A label could be used independently of a form.

* rename `form-radio-group` to just `radio-group`

A radio-group could be used independently of a form.

* rename `form-checkbox` to just `checkbox`

A checkbox could be used independently of a form.

* split "layout" and "sapcing" bundles

Moved the spacing-body classes out of utils/layout as they can be needed independently. There are cases where the layout classes aren't used, but spacing-body is still required.

* Add :host() selector to widget and elevation bundles for Shadow DOM support

Added the :host() selector for the widget and elevation bundles to allow direct application on the Host element in Shadow DOM components.

* update showcase files

* Include spacing-body tokens on accordion component.

* Update accordion component sowcase with spacing-body classes

* rename elevation classes by removing the "0"

- `.elevation-01` was renamed to `.elevation-1`
- `.elevation-02` was renamed to `.elevation-2`
- `.elevation-03` was renamed to `.elevation-3`

* add `utils/spacing` to `getThemeModelItem`

* update mercury showcase files

* Update sidebar styles and add showcase basic example.

- Add `--spacing-body-*` tokens
- Customize button styles

* remove unwanted console.log

* Include a "sidebar collapsed hidden" version

* Include "with motion" sidebar case.

- Include a "with motion" sidebar class selector.
- Update sidebar showcase.

* update showcase files

* Update all.css